requirements.txt
Flask==3.0.0
Playwright==1.40.0
requests==2.31.0
Jinja2==3.1.2
Werkzeug==3.0.1
click==8.1.7
bcrypt==4.1.1
PyJWT==2.8.1
cryptography==41.0.7
python-dotenv==1.0.0
